+++ tests/run/builtin_float.py
@@ -1,3 +1,7 @@
+# mode: run
+# tag: pure3.0
+
+import cython
 
 def empty_float():
     """
@@ -26,3 +30,31 @@ def float_call_conjugate():
     """
     x = float(1.5).conjugate()
     return x
+
+
+@cython.test_assert_path_exists(
+    "//CoerceToPyTypeNode",
+    "//CoerceToPyTypeNode//PythonCapiCallNode",
+)
+def from_bytes(s: bytes):
+    """
+    >>> from_bytes(b"123")
+    123.0
+    >>> from_bytes(b"123.25")
+    123.25
+    >>> from_bytes(b"123E100")
+    1.23e+102
+    """
+    return float(s)
+
+
+@cython.test_assert_path_exists(
+    "//CoerceToPyTypeNode",
+    "//CoerceToPyTypeNode//PythonCapiCallNode",
+)
+def from_bytes_literals():
+    """
+    >>> from_bytes_literals()
+    (123.0, 123.23, 1e+100)
+    """
+    return float(b"123"), float(b"123.23"), float(b"1e100")